requirements.txt
Flask>=2.3.0
appdirs>=1.4.4
elevenlabs>=0.2.0
ffmpeg-python>=0.2.0
google-genai>=1.0.0
mutagen>=1.47.0
openai>=1.0.0
packaging>=23.2
pillow>=10.0.0
pyinstaller>=5.13.0
pywebview>=4.4.1
replicate>=0.15.0
requests>=2.31.0
